Add files via upload
Cluster analysis plots by me and Toby. This uses hierarchical clustering and the following distances: Euclidean, Manhattan, and Minkowski p = 1.5. p = 1 and p = 2 are also included, but those are equal to Manhattan and Euclidean, respectively, we were just curious to see if they are the same or not because we saw that they should be (and they are!)